Break out Identity and User Interfaces from sdkRuntimeModels
