+++ src/apiClient.ts
@@ -1,11 +1,12 @@
 import Constants from './constants';
 import Types from './types';
 import { BatchUploader } from './batchUploader';
-import { MParticleUser, MParticleWebSDK, SDKEvent } from './sdkRuntimeModels';
+import { MParticleWebSDK, SDKEvent } from './sdkRuntimeModels';
 import KitBlocker from './kitBlocking';
 import { Dictionary, getRampNumber, isEmpty, parseNumber } from './utils';
 import { IUploadObject } from './serverModel';
 import { MPForwarder } from './forwarders.interfaces';
+import { MParticleUser } from './identity-user-interfaces';
 
 export type ForwardingStatsData = Dictionary<any>;
