+++ src/identity-utils.ts
@@ -3,7 +3,7 @@ import { Dictionary, parseNumber, isObject, generateHash } from './utils';
 import { BaseVault } from './vault';
 import Types from './types';
 import { IdentityApiData, UserIdentities, IdentityCallback } from '@mparticle/web-sdk';
-import { IdentityAPIMethod, MParticleWebSDK } from './sdkRuntimeModels';
+import { IdentityAPIMethod } from './identity.interfaces';
 
 const { Identify, Modify, Login, Logout } = Constants.IdentityMethods;
